Decimal that shit

+++ DataAccessLibrary/MarketPriceModel.cs
@@ -4,25 +4,25 @@ namespace HotelPriceScout.Data.Model
 {
     public class MarketPriceModel
     {
-        public int Price { get; set; }
+        public decimal Price { get; set; }
         public DateTime Date { get; set; }
         public string HotelName { get; set;  }
         public bool MarkedForDiscrepancy { get; set; } = false;
         public int RoomType { get; private set; } 
-        public MarketPriceModel(int price, DateTime date)
+        public MarketPriceModel(decimal price, DateTime date)
         {
             Price = price;
             Date = date;
         }
 
-        public MarketPriceModel(string hotelName, int price, DateTime date)
+        public MarketPriceModel(string hotelName, decimal price, DateTime date)
         {
             Price = price;
             Date = date;
             HotelName = hotelName;
         }
 
-        public MarketPriceModel(int price, DateTime date, int roomType)
+        public MarketPriceModel(decimal price, DateTime date, int roomType)
         {
             Price = price;
             Date = date;
